fix+refactor: tests + ConfigInfo name
