+++ src/client/types.rs
@@ -13,7 +13,7 @@ use tokio::sync::broadcast;
 
 use crate::tl::types::{
     AccountAddress, BlockId, BlockIdExt, BlocksAccountTransactionId, BlocksHeader,
-    BlocksMasterchainInfo, BlocksShards, BlocksTransactions, ChainConfigInfo, FullAccountState,
+    BlocksMasterchainInfo, BlocksShards, BlocksTransactions, ConfigInfo, FullAccountState,
     InternalTransactionId, RawFullAccountState, RawTransactions,
 };
 use crate::tl::TonNotification;
@@ -374,12 +374,12 @@ pub trait TonFunctions {
         }
     }
 
-    async fn get_config_param(&self, mode: u32, param: u32) -> anyhow::Result<ChainConfigInfo> {
+    async fn get_config_param(&self, mode: u32, param: u32) -> anyhow::Result<ConfigInfo> {
         let func = TonFunction::GetConfigParam { mode, param };
         let result = self.invoke(&func).await?;
         match result {
-            TonResult::ChainConfigInfo(result) => Ok(result),
-            r => Err(anyhow!("Expected ChainConfigInfo, got: {:?}", r)),
+            TonResult::ConfigInfo(result) => Ok(result),
+            r => Err(anyhow!("Expected ConfigInfo, got: {:?}", r)),
         }
     }
